src/test/java/fr/phoenyx/arena/controllers/item/ItemControllerTests.java
package fr.phoenyx.arena.controllers.item;

import static org.hamcrest.Matchers.containsString;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import fr.phoenyx.arena.advices.GenericAdvice;
import fr.phoenyx.arena.advices.GenericEntityAdvice;
import fr.phoenyx.arena.advices.item.ItemAdvice;
import fr.phoenyx.arena.dtos.item.ItemDTO;
import fr.phoenyx.arena.exceptions.item.ItemException;
import fr.phoenyx.arena.models.item.Item;
import fr.phoenyx.arena.services.item.ItemService;

@RunWith(SpringRunner.class)
@SpringBootTest
@AutoConfigureMockMvc
public class ItemControllerTests {

    private static final long ID = 0L;

    @Autowired
    private MockMvc mockMvc;

    @Mock
    private ItemService itemService;

    @InjectMocks
    private ItemController itemController;

    @Before
    public void setup() {
        mockMvc = MockMvcBuilders.standaloneSetup(itemController)
                .setControllerAdvice(new GenericAdvice(), new GenericEntityAdvice(), new ItemAdvice())
                .build();
    }

    @Test
    public void findAll_shouldReturnOK() throws Exception {
        mockMvc.perform(get("/items"))
                .andExpect(status().isOk());
    }

    @Test
    public void findById_shouldReturnOK_whenExists() throws Exception {
        //Given
        ItemDTO item = new ItemDTO();
        item.setId(ID);
        when(itemService.findById(ID)).thenReturn(item);

        //When Then
        mockMvc.perform(get("/items/" + ID))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(content().string(containsString(Long.toString(ID))));
    }

    @Test
    public void findById_shouldReturnNotFound_whenNotExists() throws Exception {
        //Given
        when(itemService.findById(ID)).thenThrow(ItemException.entityNotFound(ID));

        //When Then
        mockMvc.perform(get("/items/" + ID))
                .andDo(print())
                .andExpect(status().isNotFound())
                .andExpect(content().string(containsString(Item.class.getSimpleName() + " not found : " + ID)));
    }

    @Test
    public void getAllRecipes_shouldReturnOK() throws Exception {
        mockMvc.perform(get("/recipes"))
                .andExpect(status().isOk());
    }
}
